requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.4.1               # via kombu
aniso8601==4.1.0          # via flask-restful
appnope==0.1.0            # via ipython
atomicwrites==1.3.0       # via pytest
attr==0.3.1
attrs==18.2.0             # via pytest
backports.shutil-get-terminal-size==1.0.0  # via ipython
billiard==3.5.0.5         # via celery
celery==4.2.1
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via dynaconf, flask
configparser==3.7.1       # via entrypoints, flake8
decorator==4.3.2          # via ipython, traitlets
dynaconf==1.2.0
entrypoints==0.3          # via flake8
enum34==1.1.6             # via flake8, traitlets
factory-boy==2.11.1
faker==1.0.2              # via factory-boy
flake8==3.7.5
flask-jsonschema-validator==0.0.4
flask-marshmallow==0.9.0
flask-restful==0.3.7
flask-sqlalchemy==2.3.2
flask-testing==0.7.1
flask==1.0.2
freezegun==0.3.11
funcsigs==1.0.2           # via mock, pytest
functools32==3.2.3.post2 ; python_version == "2.7"
idna==2.8                 # via requests
ipaddress==1.0.22         # via faker
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==5.8.0            # via ipdb
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
jsonschema==2.6.0         # via flask-jsonschema-validator
kombu==4.3.0              # via celery
markupsafe==1.1.0         # via jinja2
marshmallow==2.18.0       # via flask-marshmallow
mccabe==0.6.1             # via flake8
mock==2.0.0
more-itertools==5.0.0     # via pytest
mysql-connector==2.1.6
pathlib2==2.3.3           # via ipython, pickleshare, pytest
pbr==5.1.2                # via mock
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.8.1             # via pytest
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.7
ptyprocess==0.6.0         # via pexpect
py==1.7.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.0           # via flake8
pygments==2.3.1           # via ipython
pytest-flask-sqlalchemy==1.0.0
pytest-mock==1.10.1       # via pytest-flask-sqlalchemy
pytest==4.2.0
python-box==3.2.4         # via dynaconf
python-dateutil==2.8.0    # via faker, freezegun
python-dotenv==0.10.1     # via dynaconf
pytz==2018.9              # via celery, flask-restful
redis==2.10.6
requests==2.21.0
scandir==1.9.0            # via pathlib2
simplegeneric==0.8.1      # via ipython
six==1.12.0               # via dynaconf, faker, flask-marshmallow, flask-restful, freezegun, mock, more-itertools, pathlib2, prompt-toolkit, pytest, python-dateutil, traitlets
sqlalchemy==1.2.17        # via flask-sqlalchemy, pytest-flask-sqlalchemy
text-unidecode==1.2       # via faker
toml==0.10.0              # via dynaconf
traitlets==4.3.2          # via ipython
typing==3.6.6             # via flake8
urllib3==1.24.1           # via requests
uwsgi==2.0.18
vine==1.2.0               # via amqp
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1          # via flask
